+++ changelog/pending/20250313--cli--remove-query-command.yaml
@@ -1,0 +1,4 @@
+changes:
+- type: chore
+  scope: cli
+  description: Remove `query` command
