+++ pkg/cmd/pulumi/org/org_search_ai.go
@@ -55,7 +55,6 @@ func (cmd *searchAICmd) Run(ctx context.Context, args []string) error {
 	displayOpts := display.Options{
 		Color:         cmdutil.GetGlobalColorization(),
 		IsInteractive: interactive,
-		Type:          display.DisplayQuery,
 	}
 	// Try to read the current project
 	ws := pkgWorkspace.Instance
