+++ pkg/cmd/pulumi/pulumi.go
@@ -65,7 +65,6 @@ import (
 	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/packagecmd"
 	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/plugin"
 	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/policy"
-	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/query"
 	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/schema"
 	cmdStack "github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/stack"
 	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/state"
@@ -411,7 +410,6 @@ func NewPulumiCmd() *cobra.Command {
 		{
 			Name: "Experimental Commands",
 			Commands: []*cobra.Command{
-				query.NewQueryCmd(),
 				convert.NewConvertCmd(),
 				operations.NewWatchCmd(),
 				logs.NewLogsCmd(),
