+++ pkg/engine/engine.go
@@ -35,16 +35,6 @@ type UpdateInfo interface {
 	GetTarget() *deploy.Target
 }
 
-// QueryInfo handles information common to query operations (list, watch).
-type QueryInfo interface {
-	// GetRoot returns the root directory for this update. This defines the scope for any filesystem resources
-	// accessed by this update.
-	GetRoot() string
-	// GetProject returns information about the project associated with this update. This includes information such as
-	// the runtime that will be used to execute the Pulumi program and the program's relative working directory.
-	GetProject() *workspace.Project
-}
-
 // Context provides cancellation, termination, and eventing options for an engine operation. It also provides
 // a way for the engine to persist snapshots, using the `SnapshotManager`.
 type Context struct {
