--- tests/integration/query/step1/Pulumi.yaml
@@ -1,3 +1,0 @@
-name: depsteps
-description: A program that inverts the dependency graph between two resources.
-runtime: nodejs
